+++ R/get_all.R
@@ -20,6 +20,8 @@ get_all <- function(job_id = "", output_dir = "./") {
   handle_setopt(h,  customrequest = "GET")
   handle_setheaders(h, "Captricity-API-Token" = Sys.getenv("CaptricityToken"))
 
-  tag_con  <- curl_fetch_memory(paste0("https://shreddr.captricity.com/api/v1/job/", job_id, "/csv/"), handle = h)
+  target_url <- paste0("https://shreddr.captricity.com/api/v1/job/", job_id,
+                       "/", "csv/")
+  tag_con  <- curl_fetch_memory(target_url, handle = h)
   curl_download(tag_con$url, destfile = paste0(output_dir, job_id, ".csv"))
 }