+++ R/test_readiness.R
@@ -12,13 +12,13 @@
 #' test_readiness("batch_id")
 #' }
 
-test_readiness <- function(batch_id="", ...) {
+test_readiness <- function(batch_id = "", ...) {
 
-  captr_CHECKAUTH()
+  if ( is.null(batch_id) | identical(batch_id, "")) {
+    stop("Provide a Valid Batch ID.")
+  }
 
-  if ( is.null(batch_id) | identical(batch_id, "")) stop("Provide a Valid Batch ID.")
-
-  res <- captr_GET(paste0("batch/", batch_id, "/readiness"), ...)
+  res <- captr_GET(paste0("batch/", batch_id, "/", "readiness"), ...)
 
   res